Fixed Issues #12, #13, #19, #29, #30, #31, #32, #33 and #34
