+++ OpenDrone-app/mobile/src/main/java/at/opendrone/opendrone/DroneCardList.java
@@ -47,18 +47,3 @@ public class DroneCardList extends DroneCardListSingleFragment {
     }
 }
 
-class Drone implements Serializable {
-
-    String name;
-    String description;
-    String type;
-
-    public Drone() {
-    }
-
-    public Drone(String name, String desc, String type) {
-        this.name = name;
-        this.description = desc;
-        this.type = type;
-    }
-}
