+++ OpenDrone-app/mobile/src/main/java/at/opendrone/opendrone/FlyStart.java
@@ -12,6 +12,7 @@ import android.support.constraint.ConstraintLayout;
 import android.support.v4.app.Fragment;
 import android.support.v4.app.FragmentManager;
 import android.support.v4.app.FragmentTransaction;
+import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -51,6 +52,7 @@ public class FlyStart extends Fragment {
         txtView_Mission = view.findViewById(R.id.txtView_Mission);
         cl_FlightPlan = view.findViewById(R.id.cl_FlightMode);
 
+        //TODO: Open the manual Fragment
         btn_FlyAuto.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
@@ -59,7 +61,7 @@ public class FlyStart extends Fragment {
 
                 switch (flightMode) {
                     case "Manual": {
-                        //TODO: Open the manual Fragment
+                        Log.i("FlyStart", "Pressed Manual");
                         break;
                     }
                     case "Automatic": {
@@ -73,10 +75,10 @@ public class FlyStart extends Fragment {
                 }
             }
         });
+        //TODO: implement TestConnection
         btn_TestConnection.setOnClickListener(new View.OnClickListener(){
             @Override
             public void onClick(View v) {
-                //TODO: implement TestConnection
                 Toast.makeText(getActivity(), "Connection established", Toast.LENGTH_SHORT).show(); //We always have a connection!
             }
         });
