+++ OpenDrone-app/mobile/src/main/res/layout/activity_drone_settings.xml
@@ -11,7 +11,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:id="@+id/ConstraintLayout"
-    tools:context=".DroneSettings"
+    tools:context=".DroneSettingsActivity"
     xmlns:app="http://schemas.android.com/apk/res-auto">
 
     <android.support.constraint.ConstraintLayout
