--- OpenDrone-app/mobile/src/main/res/layout/fragment_drone_settings.xml
@@ -1,167 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!--
-  ~ Last modified: 09.09.18 17:21
-  ~ Copyright (c) OpenDrone, 2018.  All rights reserved.
-  ~ The entire project (including this file) is licensed under the GNU GPL v3.0
-  -->
-
-<android.support.constraint.ConstraintLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:id="@+id/ConstraintLayout"
-    tools:context=".DroneSettings"
-    xmlns:app="http://schemas.android.com/apk/res-auto">
-
-    <ImageView
-        android:id="@+id/imageView_DronePicture"
-        android:layout_width="match_parent"
-        android:layout_height="150dp"
-        android:scaleType="centerCrop"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        android:src="@drawable/opendrone_bg"
-        android:contentDescription="@string/imageView_DronePicture_Description"/>
-
-    <TextView
-        android:id="@+id/text_DroneName"
-        android:layout_width="268dp"
-        android:layout_height="27dp"
-        android:layout_marginStart="15dp"
-        android:layout_marginTop="12dp"
-        android:layout_marginEnd="99dp"
-        android:textAppearance="@style/AVLoadingIndicatorView.Large"
-        android:textSize="18sp"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.535"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/imageView_DronePicture" />
-
-    <TextView
-        android:id="@+id/textView_FlightPlan"
-        android:layout_width="325dp"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="8dp"
-        android:layout_marginTop="8dp"
-        android:layout_marginEnd="8dp"
-        android:maxHeight="101dp"
-        android:text="TextView"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.0"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/text_DroneName" />
-
-    <TextView
-        android:id="@+id/txtView_Calibration"
-        android:layout_width="wrap_content"
-        android:layout_height="@dimen/txtView_default_height"
-        android:layout_marginStart="16dp"
-        android:layout_marginTop="8dp"
-        android:gravity="center_vertical"
-        android:text="@string/txtView_Calibration"
-        android:textSize="@dimen/textSize_default"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/textView_FlightPlan" />
-
-    <TextView
-        android:id="@+id/txtView_PinConfiguration"
-        android:layout_width="wrap_content"
-        android:layout_height="@dimen/txtView_default_height"
-        android:layout_marginStart="16dp"
-        android:layout_marginTop="8dp"
-        android:gravity="center_vertical"
-        android:text="@string/txtView_PinConfiguration"
-        android:textSize="@dimen/textSize_default"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/txtView_Calibration" />
-
-    <View
-        android:id="@+id/divider"
-        android:layout_width="368dp"
-        android:layout_height="1dp"
-        android:layout_marginStart="8dp"
-        android:layout_marginTop="16dp"
-        android:layout_marginEnd="8dp"
-        android:background="?android:attr/listDivider"
-        android:visibility="visible"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/txtView_PinConfiguration" />
-
-    <EditText
-        android:id="@+id/txt_DroneName"
-        android:layout_width="0dp"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="16dp"
-        android:layout_marginTop="16dp"
-        android:layout_marginEnd="8dp"
-        android:autofillHints=""
-        android:ems="10"
-        android:hint="@string/txt_DroneName"
-        android:inputType="textPersonName"
-        android:selectAllOnFocus="false"
-        android:singleLine="true"
-        android:text="@string/txt_DroneName"
-        android:textAlignment="textStart"
-        android:textColorLink="@color/primaryDarkColor"
-        android:textSize="18sp"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/divider" />
-
-    <Spinner
-        android:id="@+id/spinner_DroneTyp"
-        android:layout_width="0dp"
-        android:layout_height="@dimen/txtView_default_height"
-        android:layout_marginTop="8dp"
-        android:layout_marginEnd="16dp"
-        android:entries="@array/array_DroneTyp"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/txt_DroneName" />
-
-    <TextView
-        android:id="@+id/txtView_DroneTyp"
-        android:layout_width="0dp"
-        android:layout_height="@dimen/txtView_default_height"
-        android:layout_marginStart="16dp"
-        android:layout_marginTop="8dp"
-        android:layout_marginEnd="16dp"
-        android:gravity="center_vertical"
-        android:text="@string/txtView_DroneTyp"
-        android:textAlignment="textStart"
-        android:textSize="@dimen/textSize_default"
-        app:layout_constraintEnd_toStartOf="@+id/spinner_DroneTyp"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/txt_DroneName" />
-
-    <EditText
-        android:id="@+id/txt_Drone_description"
-        android:layout_width="0dp"
-        android:layout_height="@dimen/txtView_default_height"
-        android:layout_marginStart="16dp"
-        android:layout_marginTop="8dp"
-        android:layout_marginEnd="16dp"
-        android:ems="10"
-        android:inputType="textMultiLine"
-        android:text="@string/txt_Drone_description"
-        android:textAlignment="textStart"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/txtView_DroneTyp" />
-
-    <Button
-        android:id="@+id/btn_saveChanges"
-        style="@style/AlertDialog.AppCompat.Light"
-        android:layout_width="0dp"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="8dp"
-        android:layout_marginEnd="8dp"
-        android:layout_marginBottom="8dp"
-        android:text="@string/btn_saveChanges"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent" />
-
-</android.support.constraint.ConstraintLayout>
